#251 moved some common testing functions to utils.js
